Correct description of ``ELSE`` construct
The guide erroneously states that an ``ELSE`` construct's
 > branch is executed if "logical_expression" is true, and if none of
 > the preceding ``IF`` or ``ELSEIF`` logical conditions was true.
The ``ELSE`` construct does not have a "logical_expression" to evaluate
and it's branch is executed if none of the preceding ``IF`` or
``ELSEIF`` logical conditions are true.

Remove the erroneous statement regarding a "logical_expression"
evaluating to true in order for an ``ELSE`` construct's branch to
execute.  Also correct conjugation from "was" to "are" is second clause.